requirements.txt
alembic==1.5.8
appdirs==1.4.4
bcrypt==3.2.0
beautifulsoup4==4.9.3
blinker==1.4
bs4==0.0.1
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.4.7
cssselect==1.1.0
cssselect2==0.4.1
defusedxml==0.7.1
distlib==0.3.1
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.2
emoji==1.2.0
fake-useragent==0.1.11
Faker==7.0.1
filelock==3.0.12
Flask==1.1.2
Flask-Assets==2.0
Flask-Bootstrap==3.3.7.1
Flask-DebugToolbar==0.11.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-Moment==0.11.0
Flask-PageDown==0.3.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.5.1
Flask-WeasyPrint==0.6
Flask-WTF==0.14.3
greenlet==1.0.0
html5lib==1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jwt==1.2.0
lxml==4.6.3
Mako==1.1.4
MarkupSafe==1.1.1
numpy==1.21.0
pandas==1.3.0
parse==1.19.0
pdfkit==0.6.1
Pillow==8.1.2
psycopg2==2.8.6
pycparser==2.20
pyee==8.1.0
pyotp==2.6.0
Pyphen==0.10.0
pyppeteer==0.2.5
pyquery==1.4.3
python-dateutil==2.8.1
python-dotenv==0.16.0
python-editor==1.0.4
pytz==2021.1
qrcode==7.1
requests==2.25.1
requests-html==0.10.0
selenium==3.141.0
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.4.4
text-unidecode==1.3
tinycss2==1.1.0
tqdm==4.61.1
urllib3==1.26.4
virtualenv==20.4.3
virtualenvwrapper-win==1.2.6
visitor==0.1.3
w3lib==1.22.0
WeasyPrint==52.5
webassets==2.0
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
WTForms==2.3.3
